+++ src/component/mod.rs
@@ -1,5 +1,5 @@
 pub mod day;
-pub mod day_breakdown;
+pub mod day_breakdown_chart;
 pub mod interval_view;
 pub mod log_table;
 pub mod token_summary_table;
